added not working kanban scrum page
