+++ spacebar/package.json
@@ -22,6 +22,7 @@
     "react-scripts": "4.0.3",
     "react-scroll": "^1.8.2",
     "react-swipeable-views": "^0.14.0",
+    "recharts": "^2.0.9",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
